drfecommerce/requirements.txt
asgiref==3.8.1
attrs==25.1.0
coverage==7.6.12
Django==5.1.6
django-js-asset==3.1.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==36.1.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-factoryboy==2.7.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
